Refactored to all.madara Included other manga websites:  - mangasushi.net  - ninjascans.com  - readmanhua.net  - zeroscans.com
