+++ src/all/madara/build.gradle
@@ -2,9 +2,9 @@ apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 
 ext {
-    appName = 'Tachiyomi: LeviatanScans'
-    pkgNameSuffix = "en.leviatanscans"
-    extClass = '.LeviatanScansFactory'
+    appName = 'Tachiyomi: Madara'
+    pkgNameSuffix = "all.madara"
+    extClass = '.MadaraFactory'
     extVersionCode = 1
     libVersion = '1.2'
 }
