+++ src/all/madara/res/mipmap-hdpi/ic_launcher.png
