+++ src/all/madara/res/mipmap-mdpi/ic_launcher.png
