+++ src/all/madara/res/mipmap-xhdpi/ic_launcher.png
