+++ src/all/madara/res/mipmap-xxhdpi/ic_launcher.png
