+++ src/all/madara/res/mipmap-xxxhdpi/ic_launcher.png
